Create feature module in feature_extractor
Moved there compute_feature from __init__.py
